refactor(template): push client boundaries to leaf components in account sidebar and tabs
Move "use client" from file-level sidebar.tsx and mobile-tabs.tsx to tiny
leaf components that handle only the interactive part (usePathname for active
state). The presentational shell (7 sidebar + 2 tab components) is now
server-rendered, reducing client JS and resolving translations on the server.
